system_status_monitoring/requirements.txt
APScheduler==3.9.1.post1
coreapi==2.3.3
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
mysqlclient==2.1.1
psutil==5.9.4
pygal==3.0.0
